\name{check_repr}
\alias{check_repr}
\title{Check representation}
\usage{
  check_repr(x)
}
\arguments{
  \item{x}{Number}
}
\description{
  The function returns the amount of bytes that is required
  to store a specific number. It is primarily used by the
  `put_variable`-method in the cdb class.
}

