\docType{methods}
\name{doc}
\alias{doc}
\alias{initialize}
\title{Coldbir doc class}
\usage{
  doc(...)
}
\arguments{
  \item{...}{objects passed on to a list(...)}
}
\description{
  Coldbir documentation class

  Method that runs when a new doc object is initialized.

  Function to create a new Coldbir documentation object
}
\examples{
r <- doc(title = "GDP", description = "Gross domestic product")
}

