\docType{methods}
\name{`[`}
\alias{`[`}
\title{Extract content from variable}
\arguments{
  \item{x}{cdb object}

  \item{i}{variable name}

  \item{j}{variable dims}

  \item{na}{NA value (default: NA)}
}
\description{
  Function to extract data content from a Coldbir variable.
  Overloads the `[`-method.
}

