\name{get_documentation}
\alias{get_documentation}
\title{Get documentation from disk}
\usage{
  get_documentation(name, path, console = FALSE,
    file_name = .doc_json)
}
\arguments{
  \item{name}{Variable name}

  \item{path}{Directory of where the file are to be
  created}

  \item{console}{TRUE if print to console instead of to
  character string}

  \item{file_name}{Documentation file name}
}
\description{
  Read documentation of a variable from disk.
}

