\name{get_readme}
\alias{get_readme}
\title{Get readme from disk}
\usage{
  get_readme(name, path, console = FALSE)
}
\arguments{
  \item{name}{Variable name}

  \item{path}{Directory of where the file are to be
  created}

  \item{console}{TRUE if print to console instead of to
  character string}
}
\description{
  Read readme of a variable from disk.
}

