\name{get_variable_doc}
\alias{get_variable_doc}
\title{Get variable documentation from disk}
\usage{
  get_variable_doc(name, path, console = FALSE,
    file_name = .doc_file)
}
\arguments{
  \item{name}{Variable name}

  \item{path}{Directory of where the file are to be
  created}

  \item{console}{TRUE if print to console instead of to
  character string}

  \item{file_name}{Documentation file name}
}
\description{
  Read documentation of a variable from disk.
}

