\name{key_to_value}
\alias{key_to_value}
\title{Replace keys with values in vector}
\usage{
  key_to_value(v, ht, factors = FALSE)
}
\arguments{
  \item{v}{Vector}

  \item{ht}{Hash table}

  \item{factors}{If vector should be returned as factors,
  default is FALSE.}
}
\description{
  Function to replace keys with values in a variable vector
}

