\name{list_match}
\alias{list_match}
\title{Match two lists}
\usage{
  list_match(x, val)
}
\arguments{
  \item{x}{data list, including all database variables}

  \item{val}{matching list, could also include wild cards
  (._)}
}
\description{
  See what part of a list that is also included in another
  one. Returns the inner join of both lists.
}

