\name{list_to_md}
\alias{list_to_md}
\title{Convert list to markdown}
\usage{
  list_to_md(lst, str = "", h = 1)
}
\arguments{
  \item{lst}{List to convert}

  \item{str}{Return string}

  \item{h}{Header start level}
}
\description{
  Function to convert a list to a character string
  containing markdown syntax
}

