\name{list_variables}
\alias{list_variables}
\title{List all variables}
\usage{
  list_variables(path = getwd(), dims = F, ...)
}
\arguments{
  \item{path}{Path}

  \item{dims}{If dims should be included (returns a data
  table)}

  \item{...}{search_files() arguments}
}
\description{
  Get a list of all variables (including dimensions)
  available in the database.
}

