\name{put_keys}
\alias{put_keys}
\title{Write keys to disk}
\usage{
  put_keys(df, name, path = getwd(), create_dir = TRUE)
}
\arguments{
  \item{df}{Two-column data frame with keys and values}

  \item{name}{Variable name}

  \item{path}{Directory of where the file are to be
  created}

  \item{create_dir}{If folder should be created when
  missing}
}
\description{
  Write keys that represents variable data to disk.
}

