\name{put_readme}
\alias{put_readme}
\title{Put readme to disk}
\usage{
  put_readme(lst, name, path = getwd(), create_dir = TRUE)
}
\arguments{
  \item{lst}{Readme list}

  \item{name}{Variable name}

  \item{path}{Directory of where the file are to be
  created}

  \item{create_dir}{If folder should be created when
  missing}
}
\description{
  Write readme of a variable to disk.
}

