\docType{methods}
\name{`[<-`}
\alias{`[<-`}
\title{Assign content to variable}
\arguments{
  \item{x}{cdb object}

  \item{i}{variable name}

  \item{j}{variable dims}

  \item{value}{value}
}
\description{
  Function to assign content, either data or documentation,
  to a Coldbir variable. Overloads the `[<-`-method.
}

