\name{file_path}
\alias{file_path}
\title{Generate coldbir file path}
\usage{
  file_path(name, path, dims = NULL, ext,
    create_dir = FALSE)
}
\arguments{
  \item{name}{Variable name}

  \item{path}{Variable path}

  \item{dims}{Vector specifying the dimensions}

  \item{ext}{File extension}

  \item{create_dir}{}
}
\description{
  ...
}

