\name{file_path}
\alias{file_path}
\title{Generate coldbir file path}
\usage{
  file_path(name, path, dims = NULL, ext = NULL,
    create_dir = F, file_name = T, data_folder = T)
}
\arguments{
  \item{name}{Variable name}

  \item{path}{Variable path}

  \item{dims}{Vector specifying the dimensions}

  \item{ext}{File extension}

  \item{create_dir}{If folder should be created when
  missing}

  \item{file_name}{If file name should be included in
  return path}

  \item{data_folder}{If data folder should be added}
}
\description{
  ...
}

