\name{put_db}
\alias{put_db}
\title{Write data frame columns to disk}
\usage{
  put_db(df, path = getwd(), dicts = TRUE)
}
\arguments{
  \item{df}{Data frame}

  \item{path}{Database path}

  \item{dicts}{If dictionaries should be created}
}
\description{
  Function to write all columns of a data frame to disk.
}

