\name{put_readme}
\alias{put_readme}
\title{Put readme to disk}
\usage{
  put_readme(name, path = getwd(), create_dir = TRUE, ...)
}
\arguments{
  \item{name}{Variable name}

  \item{path}{Directory of where the file are to be
  created}

  \item{create_dir}{If folder should be created when
  missing}

  \item{...}{list_to_md(...)}
}
\description{
  Write readme of a variable to disk.
}

