\name{recursive_list}
\alias{recursive_list}
\title{Create recursive list from vector}
\usage{
  recursive_list(x, val)
}
\arguments{
  \item{x}{character vector}

  \item{val}{endpoint value}
}
\description{
  A helper function to convert variable file names to a
  list representation
}
\examples{
\dontrun{
x <- c("a", "b", "c")
recursive_list(x)
}
}

