\name{to_values}
\alias{to_values}
\title{Replace keys with values in vector}
\usage{
  to_values(x, ..., factors = FALSE)
}
\arguments{
  \item{x}{Vector}

  \item{...}{See get_lookup().}

  \item{factors}{If vector should be returned as factors,
  default is FALSE.}
}
\description{
  Function to replace keys with values in a variable vector
}

